+++ runtime-k3d/src/app-deployment/config/podspec/vehicleapp.yaml
