+++ runtime-k3d/src/app-deployment/deploy_vehicleapp.py
@@ -122,7 +122,6 @@ def deploy_vehicleapp():
             else:
                 spinner.write(f"{status} vapp-chart not yet installed.")
 
-            spinner.write(status)
             install_vehicleapp(app_name)
             spinner.write(f"> Installing vapp-chart for {app_name}... done!")
             spinner.ok("✔")
