+++ runtime-k3d/src/runtime/deployment/config/helm/templates/bash.yaml
